Extract the marketType
