+++ src/constants/markets.ts
@@ -1,3 +1,2 @@
 export const XBTUSD = 'PI_XBTUSD';
 export const ETHUSD = 'PI_ETHUSD';
-export type marketType = typeof XBTUSD | typeof ETHUSD;
