+++ src/types/market.ts
@@ -1,0 +1,3 @@
+import { XBTUSD, ETHUSD } from '../constants/markets';
+
+export type marketType = typeof XBTUSD | typeof ETHUSD;
